refac: tests to mod.rs
